+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Let class-transformer deserialize arrays directly
 * Add cache summary to `getCurrentState()`'s storage info
 
 ## v0.65.0